moved the rollup sized plugin into a plugins directory
